+++ spring-cloud-zookeeper-discovery/src/main/java/org/springframework/cloud/zookeeper/discovery/dependency/ZookeeperDependencies.java
@@ -75,7 +75,7 @@ public class ZookeeperDependencies {
 		return !dependencies.isEmpty();
 	}
 
-	public ZookeeperDependency getDependencyForPath(String path) {
+	public ZookeeperDependency getDependencyForPath(final String path) {
 		for (Map.Entry<String, ZookeeperDependency> zookeeperDependencyEntry : dependencies.entrySet()) {
 			if (zookeeperDependencyEntry.getValue().getPath().equals(path)) {
 				return zookeeperDependencyEntry.getValue();
