+++ spring-cloud-zookeeper-discovery/src/test/groovy/org/springframework/cloud/zookeeper/discovery/dependency/ZookeeperDependenciesSpec.groovy
@@ -25,29 +25,29 @@ class ZookeeperDependenciesSpec extends Specification {
 		expect:
 			exptectedDependency == zookeeperDependencies.getDependencyForPath(path)
 		where:
-			path   || exptectedDependency
-			null   || null
-			'path' || EXPECTED_DEPENDENCY
+			path          || exptectedDependency
+			'unknownPath' || null
+			'path'        || EXPECTED_DEPENDENCY
 	}
 
 	@Unroll
-	def "should retrieve dependency [#alias] for alias [#alias]"() {
+	def "should retrieve dependency [#exptectedDependency] for alias [#alias]"() {
 		expect:
 			exptectedDependency == zookeeperDependencies.getDependencyForAlias(alias)
 		where:
-			alias   || exptectedDependency
-			null    || null
-			'alias' || EXPECTED_DEPENDENCY
+			alias          || exptectedDependency
+			'unknownAlias' || null
+			'alias'        || EXPECTED_DEPENDENCY
 	}
 
 	@Unroll
-	def "should retrieve alias [#alias] for path [#path]"() {
+	def "should retrieve alias [#expectedAlias] for path [#path]"() {
 		expect:
 			expectedAlias == zookeeperDependencies.getAliasForPath(path)
 		where:
-			path   || expectedAlias
-			null   || ''
-			'path' || 'alias'
+			path          || expectedAlias
+			'unknownPath' || ''
+			'path'        || 'alias'
 	}
 
 	@Unroll
@@ -55,9 +55,9 @@ class ZookeeperDependenciesSpec extends Specification {
 		expect:
 			expectedPath == zookeeperDependencies.getPathForAlias(alias)
 		where:
-			alias   || expectedPath
-			null    || ''
-			'alias' || 'path'
+			alias          || expectedPath
+			'unknownAlias' || ''
+			'alias'        || 'path'
 	}
