+++ spring-cloud-zookeeper-discovery/src/test/groovy/org/springframework/cloud/zookeeper/discovery/dependency/ZookeeperDiscoveryWithDependenciesISpec.groovy
@@ -23,9 +23,9 @@ import org.springframework.cloud.client.ServiceInstance
 import org.springframework.cloud.client.discovery.DiscoveryClient
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient
 import org.springframework.cloud.client.loadbalancer.LoadBalanced
-import org.springframework.cloud.zookeeper.config.CommonTestConfig
+import org.springframework.cloud.zookeeper.common.CommonTestConfig
+import org.springframework.cloud.zookeeper.common.TestRibbonClient
 import org.springframework.cloud.zookeeper.discovery.PollingUtils
-import org.springframework.cloud.zookeeper.discovery.TestServiceRestClient
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
@@ -103,13 +103,6 @@ class ZookeeperDiscoveryWithDependenciesISpec extends Specification implements P
 
 		@RequestMapping('/ping') String ping() {
 			return 'pong'
-		}
-	}
-
-	static class TestRibbonClient extends TestServiceRestClient {
-
-		TestRibbonClient(RestTemplate restTemplate) {
-			super(restTemplate)
 		}
 	}
 }
